better abortcontroller handling and now we disconnect wherever we click when we are in a game correctly
